Initial implementation of a Vector Graphic HD/FD Disk Controller with Wishbone interface.  See the PDF in /doc for more information about this controller.
For now, the storage is implemented in FLASH memory, which is read-only
to the controller.  Also, the controller does not check (CRC) the disk
sector data, like the original does.  Instead, it inserts 00's where the
CRC goes, to simulate a correctly checked CRC.

I am able to boot CP/M 2.2 using this controller, with the vgboot.mcs image
programmed into FLASH.  See the diskimage/ directory.